エージェント部分をリファクタリングした。レガシーなコードを分離型に変更した。
