+++ back/log/ManimGraphAnimationService.log
@@ -269,3 +269,9 @@ code: │   axes_labels = axes.get_x_axis().add_labels(labels)                  
 2025-11-03 09:23:38.246 | INFO     | app.service.graph_agent:_execute_and_handle_errors:123 -    [+] Script executed successfully.
 2025-11-03 09:23:38.246 | INFO     | app.service.graph_agent:_after_execution:227 - --- [Branch] Execution Succeeded. Ending Graph. ---
 2025-11-03 09:23:38.246 | INFO     | app.service.graph_agent:generate_video:278 - --- Graph Finished: Success ---
+2025-11-03 09:52:43.670 | INFO     | app.service.base_agent:_setup_logger:78 - Logger initialized. Log file: log/ManimGraphAnimationService.log
+2025-11-03 09:52:43.671 | INFO     | app.service.base_agent:_load_prompt:93 - Prompts loaded from /workspaces/ai_agent/back/app/service/prompt/prompts.toml
+2025-11-03 09:54:02.183 | INFO     | app.service.base_agent:_setup_logger:78 - Logger initialized. Log file: log/ManimRegacyAgentService.log
+2025-11-03 09:54:02.184 | INFO     | app.service.base_agent:_load_prompt:93 - Prompts loaded from /workspaces/ai_agent/back/app/service/prompt/prompts.toml
+2025-11-03 09:55:38.416 | INFO     | app.service.base_agent:_save_script:129 - Script saved to tmp/regacy_agent.py
+2025-11-03 09:56:11.448 | INFO     | app.service.base_agent:_execute_script:153 - Script executed successfully: tmp/regacy_agent.py
